requirements.txt
python==3.10
streamlit==1.25.0
opencv-python==4.8.0.76
Pillow==9.5.0
numpy==1.23.5
ultralytics==8.0.0  # YOLO model library
